+++ GlobalProcessingLibrary/Source/ModelADA.cpp
@@ -111,7 +111,7 @@ double d_kappa_d_tau(double tau2, double tau1)
    return d;
 }
 
-void updatestatus_(int* gc_int, int* thread, int* iter, double* chi2, int* terminate)
+void updatestatus_(int* gc_int, int* thread, int* iter, float* chi2, int* terminate)
 {
    FLIMGlobalFitController* gc= (FLIMGlobalFitController*) gc_int;
    int t = gc->status->UpdateStatus(*thread, -1, *iter, *chi2);
